Test trivial padding returns the original array
The `da.pad` function handles padding of the ends by using
`da.concatenate` to join the edges with the original array. Thus if the
arrays to pad were trivial in size, they would be joined as well.
However this results in some unnecessary changes to the graph.

The `da.concatenate` function has since been rewritten to drop arrays of
trivial length from the result. Thus `da.pad` benefits from this
behavior and will also exclude arrays of trivial length. Though this was
not tested previously. So here we include a test to ensure `da.pad`
doesn't change the array when trivial padding occurs.