Add recompute= keyword to svd_compressed for lower-memory use (#5041)
The svd_compressed algorithm needs to use the input array a few times
Often, this array is too large to fit into memory, and so it is better
to recompute it.  This PR adds a keyword, ``compute=``, which
optionally triggers this behavior.  It is turned off by default.

We also fix up the docstring a bit and one of the keys in the high level
graph.

This also introduces a dask.distributed agnostic wait function to wait
on persisted data if we're using the distributed scheduler, but to do
nothing otherwise.

With this change we're able to compute arbitrarily large svd
computations in low memory.