Use da.from_array(..., asarray=False) if the input implements __array_function__
Previously we called `np.asarray` on each chunk in `da.from_array`.
This can be troublesome if we have arrays like `cupy` or `sparse` that
are capable of performing computations on their own.

Now we look for an `__array_function__` method on the input as a signal
that it can handle numpy-like functions.  If it has this method then we
don't call `np.asarray`