Add alternative SSHCluster implementation (#2827)
This is a proof of concept here for two reasons:

1.  It opens up a possible alternative for SSH deployment (which was
    surprisingly popular in the user survey)
2.  It is the first non-local application of `SpecCluster` and so serves
    as a proof of concept for other future deployments that are mostly
    defined by creating a remote Worker/Scheduler object

This forced some changes in `SpecCluster`, notably we now have an `rpc`
object that does remote calls rather than accessing the scheduler
directly.  Also, we're going to have to figure out how to handle all of
the keyword arguments.  In this case we need to pass them from Python
down to the CLI, and presumably we'll also want a `dask-ssh` CLI command
which has to translate the other way.