Close workers gracefully with --lifetime keywords (#2892)
This allows workers to optionally terminate themselves gracefully after a predetermined time.  This can be helpful in a few contexts:

1.  We receive a SIGINT, and know that we need to clean up quickly (though note that the signal handlers are not implemented as part of this commit
2.  We know that we'll be kicked off at a certain time, such as in one hour from now, as is often specified by HPC job schedulers 
3.  We just want to refresh our workers every once in a while, because we know that our code leaks some memory .  Fixes https://github.com/dask/distributed/issues/2861 

This is configurable as keywords to the `Worker` or `Nanny` classes, in config values, or with CLI.  Here is an example with CLI.

### Restart to clear state

```
dask-worker scheduler:8786 --lifetime 1hr --lifetime-restart --lifetime-stagger 5m
```

This will kill the worker roughly 1 hour from now +- a range of 5 minutes (to avoid killing all of our workers at the same time).  It will also allow that worker to be restarted afterwards

### Restart to avoid walltime death

```
dask-worker scheduler:8786 --lifetime 58m 
```

Here we don't try to restart the worker (no point) and we choose a time a bit before our 60m walltime.