fix(dataframe): allow append mode in to_csv()
Previously the mode argument of DataFrame.to_csv() is ignored, and the
file mode is hardcoded to write mode. This prevents appending data
frames to an existing CSV file.

Accept the mode argument and use the previously hardcoded value as its
default. As martindurant pointed out, not all filesystems support append
mode.

Fix #5187.