A few changes to ``config.set``
The main intent of this PR is to support setting config values through
keyword parameters instead of requiring a dictionary for nested config
values. We follow the same convention as we do for environment
variables, where double-underscores are interpreted as a ``.``, and
indicate an increase in nesting level. Now the following are equivalent:

```python
with dask.config.set({"foo.bar.baz": 1}):
    ...

with dask.config.set(foo__bar__baz=1):
    ...
```

If a mapping argument and keyword arguments are both provided, the
mapping is applied first, then the keyword arguments (the keywords are
effectively overrides for the mapping).

In the process, I discovered a bug in the rollback code when dealing
with multiple keys interacting with the same nested value. To deal with
this I rewrote the rollback code in a way that was simpler for me to
understand, and hopefully others as well.