Update the use of fixtures in distributed tests (#5497)
Our CI was failing with the 2.6 release.  This fixes a bug there, and
also modernizes how we test to be more in line with the dask.distributed
library.

Also, the pre-commit commit for black was giving me trouble.  I've
downgraded this to the commit that we use in distributed, which seems to
work well.