Fixed independent optimization of parquet-getitem
Fixed independent optimization of parquet-getitem

On master, the parquet-getitem optimization can cause us to form
invalid Dask graphs, as we (somehow I don't fully understand) reuse an
optimized `read_parquet` that doesn't contain all the keys needed for a
secondary `__getitem__`.

This fixes that by changing the `name` used for the `read_parquet`, and
all references to it.

Closes https://github.com/dask/dask/issues/5502