Make Listener.start asynchronous (#3278)
Previously the Listener.start method was synchronous,
expecting comms to be able to start themselves and get an address
immediately.  This was a challenge when we tried adding asyncio comms,
which needed to await the server creation.

Now we make Listener.start asynchronous, and make sure that we await all
calls to it within the codebase.