Update `dask.order.order` to consider "next" nodes using both FIFO and LIFO.
Closes #5859.  @JSKenyon

How do we choose the next node in `order` after a node with no dependents?
The answer is from one of dependents of an already completed node.
However, this is difficult to do well while also keeping things fast.
Previously, we chose dependents by considering completed nodes in a FIFO manner.
Now, we consider both LIFO and FIFO and choose the best ones.
This seems to do reasonably well and is fast enough.