Support serializing/deserializing `rmm.DeviceBuffer`s (#3442)
* Serialize and deserialize RMM `DeviceBuffer`'s

* Register RMM serializers and deserializers

* Test RMM serialization/deserialization

* Test serializing RMM `DeviceBuffer` from Numba

* Test deserializing a CuPy array from RMM data

* Test deserializing a Numba array from RMM data

* Fix some minor formatting issues

* Drop unneeded `cuda.as_cuda_array` call

This should already be deserialized as a Numba array. So go ahead and
treat it as such.

* Use namespace with `DeviceBuffer`

* Assert that `arr` is a `rmm.DeviceBuffer`

If we got here, we should already have an `rmm.DeviceBuffer` as that is
what we would have allocated. If that's not the case, something very
wrong has happened. So just `assert` that is true (erroring otherwise).

* Drop unneeded Numba-backed RMM serialization test

As RMM is used preferentially for allocations when available, there
shouldn't be a case where we need to serialize an RMM `DeviceBuffer`
from a Numba array. So drop this test.

* Drop unused import