Better layer name synthesis for non-HLGs
When a dependent collection passed to HighLevelGraph.from_collections
has a __dask_graph__ that is not a HighLevelGraph, the layer synthesised
for it was being named with `id()`. This caused problems, because there
are places in the code that assume that the layer name matches the key
names inside the layer (see #5850), and dask.array.utils._check_dsk
asserts that layer names are all tuples or strings, not integers.

Address withby naming the layer `collection.__dask_layers__()[0]`,
falling back to the old behaviour if `__dask_layers__` doesn't exist or
returns an empty tuple.  I'm not sure what the ideal behaviour is if
`__dask_layers__()` returns more than one element. The built-in
collections (array, dataframe, bag and delayed) all return a singleton;
xarray's Dataset can return multiple values, but it also returns a
HighLevelGraph from `__dask_graph__` so the issue doesn't arise.

Closes #5850.