Push chunk override from map_blocks into blockwise
Instead of calling dask.array.blockwise and then fiddling with the
chunks under the hood, pass the new chunking scheme into blockwise.
The verification that the number of blocks didn't change is now made
inside blockwise, so it will be a bit more robust.

To make that verification work I also improved the code inside blockwise
that estimates the chunks up front (in the adjust_chunks=False case, as
used by map_blocks). Previously it made the arbitrary choice of taking
the input with the most indices. When there is broadcasting, that can
get the wrong *number* of blocks (which is what lead to bugs to
like #4299). It now, per dimension, takes the chunk scheme with the most
blocks. Note that this might alter the behaviour in cases where the
chunking schemes are inconsistent and no explicit chunking scheme is
provided, but I believe the behaviour was undefined to start with.

A benefit of this change is that it's now possible to detect cases where
the user-provided chunks have more than one element and there should
only be more than one element (previously the check was skipped because
broadcasting wasn't correctly handled, as a workaround for #4299).