Make map_blocks with block_info produce a Blockwise
Previously, using map_blocks with block_info or block_id would denature
a Blockwise and do a substitution on every instantiation of the subgraph
callable to inject the block info.

Instead, create a dask array whose elements are the block infos, and
pass this as an extra parameter into `dask.array.blockwise`. This makes
the layer a Blockwise, and hence a candidate for `rewrite_blockwise`.