Serialize 1-D, contiguous, `uint8` CUDA frames (#3475)
* Space things out to improve readability

* Assign `frames` for clarity

* Check CUDA contiguous arrays the same way

* Use keyword arguments with CUDA array constructors

Should make it a little clearer what the arguments relate to. Also makes
them less dependent on changes in the signature of the constructors.

* Always serialize the CUDA array's `strides`

* Make sure CuPy always uses the `strides` provided

* Drop workarounds for unknown `strides`

As we always include the actual `strides` in the `header`, we can rely
on this when deserializing the data. So drop the workarounds added for
C-contiguous data where `strides` could be `None` as we still have the
exact `strides` in that case.

* Convert CUDA arrays into 1-D contiguous arrays

* Avoid copying C/F contiguous CUDA arrays

* Cast CUDA array data to `uint8` before serializing

As we will ultimately read any data in as `uint8`, go ahead and cast to
`uint8` as part of serialization. This is a good first order check to
make sure that we are able to serialize the dat

To simplify handling of the data a bit, go ahead and cast it to `uint8`
before serializing. This makes contiguity checks trivial.

* Test serialization of F-contiguous CUDA arrays