Go big!  Make `order` smarter about when it chooses (or changes!) tactical goals.
The old old method used to switch tactical goals frequently.  Current master
does not switch tactical goals at all.  This commit compromises and switches
tactical goals when it is obviously better.

Also, instead of trying to do a LIFO or FIFO or LIFO/FIFO policy for choosing
what to compute after a tactical goal is reached, we now have a policy that
allows continuing from any node (not just the last or first).  In order for
this to behave well for a variety of use cases, we needed to introduce the
concept of "epochs" and "eras", which define what can be computed later.

Performance-wise, this remains pretty good.  It is often a little slower, but
not significantly so.  We avoided pathological scaling behaviors.