Add dask serialization of CUDA objects (#3482)
* Run `isort` on CUDA protocol `import`s

* Align CuPy serialize/deserialize function names

* Prefix CUDA serializers with `cuda_`

This should make room for Dask serializers to also be specified and
added.

* Add Dask serializers for RMM `DeviceBuffer`s

To make TCP a bit more performant with RMM, provide Dask serializers to
allow going to and from host memory.

* Add Dask serializers for Numba `DeviceNDArray`s

* Add Dask serializers for CuPy `ndarray`s

* Parametrize serializers in CUDA object tests

To make sure that different CUDA objects can use different serialization
protocols, test with each one individual and ensure it completes. In
particular test both "cuda" and "dask". Where supported also test
"pickle", but skip it when it is not (like with Numba).

* Check frames are the expected type

To make sure Dask can handle transmission of the frames serialized, test
they match the type expected by the protocol used. With "cuda" ensure we
get something that supports `__cuda_array_interface__`. With "dask" make
sure we get a `memoryview`.