Pin `numpydoc` to `0.8.0` (fix double autoescape) (#5961)
This is a temporary fix for the doubly escaped `*`s that are present in
the current dask docs (and `pandas`, and `scipy`...).
This issue started with `numpydoc=0.9.0` when `numpydoc` started
escaping `*args` and `**kwargs`.  sphinx autosummary _also_ escapes
`*args` and `**kwargs`, so in conjunction, they end up double-escaped:

```
Client.run(self, function, \*args, \*\*kwargs)
```

There are a couple of different ways to fix this TheRightWay™ but this
will make the docs prettier for now.