Yet another revision to `order`.
This applies a few of the lessons we have learned recently:
  1. there needs to be a way to change path along dependents if a better path is found.
  2. there needs to be a robust way to decide what to do next after a tactical goal
     (node with no dependents) is reached.
  3. when deciding what to do next, we should ideally take into account how close
     a node is to being completed (i.e., not just consider the final goal).

To do all these, we partition dependents of a node based on a score that indicates how
much future work is associated with the node.  We use this score to determine whether
tactical goals should change.  When all tactical goals are reached (so `inner_stack`
and `inner_stacks` are empty), we continue with the dependents of completed nodes that
had the lowest partition score.  We keep performance respectable by (hopefully)
partitioning nodes into many small groups (recall that it's much better to sort many
small groups than one large group).