Support using other serializers with `register_generic` (#3536)
* Build `ObjectDictSerializer`

Makes a class to handle serialization and deserialization of objects
with `__dict__`s. Should make it easier to specify other serializers one
can use with such objects.

Note if we are unable to use the serializer specified, we fallback to
pickling things.

* Use `ObjectDictSerializer` for "dask" serialization

* Make `register_generic` handle other serializers

Allow users to call `register_generic` with other serializers in mind.
By default still use `dask_serialize` and `dask_deserialize`. Though
allow these to be subbed out by other `*_serialize` and `*_deserialize`
functions that will also dispatch based on type to serialize and
deserialize with the expected serialization mode.

* Ensure `dict` can deserialized with "cuda"

As nested serialization can run into `dict` objects in the process, make
sure that they can be deserialized with "cuda" as well to produce the
original object.