Improve reuse of temporaries with numpy (#5933)
numpy has an optimisation that allows operations to be done in-place if
an operand is a temporary (has no more references) e.g. in the
expression `a + b + c`, `c` will be added in-place to the result of `a +
b`.

Tweak the signle-threaded scheduler to avoid creating unnecessary
references. To make this effective, also modify Blockwise to run
`fuse` on the subgraph before creating SubgraphCallable.