Check `nbytes` and `types` before reading `data` (#3628)
* Check `nbytes` and `types` before reading `data`

To avoid reading `data` when it is not needed, try checking `nbytes` and
`types` beforehand. If the metadata is already there, continue on
without reading `data`. Otherwise fallback to the reading `data`, but do
make sure to cache the results of that read to avoid doing it again.

* Use `.get(...)` with `self.nbytes` as well

* Restart GitHub CI

Appears GitHub CI failed to checkout the code and clicking restart in
the UI does not work. So pushing a dummy commit to restart it.