Add Semaphore extension (#3573)
The complexity of the internal structure comes in since we do not support any notion of an ephemeral key, i.e. a value which expires together with the session. In this context this translates best to the Client. Therefore, the Semaphore tracks which lease stems from which Semaphore client instance and stores its associated Client ID. If the client is lost/closed, the semaphore will release all it's acquired values eventually.

This behavior is quite important for resilience: If a worker is shut down ungracefully, all acquired leases should be released eventually since otherwise we may cause a deadlock.

gh3573 and gh2690