Get CUDA context to finalize Numba `DeviceNDArray` (#3666)
* Ensure CUDA context on `DeviceNDArray` cleanup

As the CUDA context does not seem to always be established when cleaning
up Numba `DeviceNDArray`s, make sure to get the context right before
cleanup to ensure the context is available.

This is done when allocating new CUDA frames to when using UCX. Make
sure this is handled both for Numba and old versions of RMM that also
use Numba under-the-hood. Newer versions of RMM don't need this.

Also do this when performing a host-to-device transfer where a Numba
`DeviceNDArray` is the first object created on device.

* Drop `PatchedCudaArrayInterface` and usage thereof

This was used to ensure that Numba had acquired a CUDA context before
finalizing any Numba `DeviceNDArray` objects that might be used to back
a CuPy `ndarray`. As we now make sure of this when creating Numba
`DeviceNDArray` by using `weakref.finalize`, this should be correctly
handled for all CUDA objects backed by Numba `DeviceNDArray`s (not just
CuPy `ndarray`s). So there should be no need to do this for CuPy
`ndarray`s separately. Should simplify the CuPy serialization and
cleanup paths a bit.