Improve `order` by putting more dependents on inner_stacks when it's safe too.
Previously, `order` only added one dependent to `inner_stack` when `inner_stack` is empty.
Now, it adds all the other dependents with the minimum partition key to `inner_stacks`.
Prior to this commit, the new changes did a lot better with this example

https://github.com/dask/dask/issues/5859#issuecomment-608422198

but it didn't quite pass the new test.  In general, this means that non-optimal behavior
can happen when `inner_stacks` gets processed while there are better options available
in `next_nodes`.  Typically, this should only result in a modest delay.  We could do
better if we pushed dependents into, say, a sorted heapq, but this would kill performance.