Replace Bokeh Server with Tornado HTTPServer (#3658)
This creates a vanilla Tornado HTTPServer on the Worker and Schedulers, and then optionally attaches a Bokeh application to that server.

This lets us always have an HTTP Server running, even if we aren't running the bokeh dashboard. This is helpful for example for prometheus and health checks, and to allow for other uses of the HTTP Server in the future. I hope that it also makes it easier to configure HTTPS globally across the project.

The info, json, proxy, health, and prometheus routes have been moved out of the dashboard directory into a new distributed/http directory.