Update `dask.order.order` to consider "next" nodes using both FIFO and LIFO (#5872)
* Update `dask.order.order` to consider "next" nodes using both FIFO and LIFO.

Closes #5859.  @JSKenyon

How do we choose the next node in `order` after a node with no dependents?
The answer is from one of dependents of an already completed node.
However, this is difficult to do well while also keeping things fast.
Previously, we chose dependents by considering completed nodes in a FIFO manner.
Now, we consider both LIFO and FIFO and choose the best ones.
This seems to do reasonably well and is fast enough.

* Hmm, let's try something else.  Need to update comments.

When repopulating `outer_stack`, consider and use all dependents of all completed tasks.
Partition them based on a simple metric, then trust FIFO ordering, so no need to sort.
This remains fast.  Fingers crossed!

* Go big!  Make `order` smarter about when it chooses (or changes!) tactical goals.

The old old method used to switch tactical goals frequently.  Current master
does not switch tactical goals at all.  This commit compromises and switches
tactical goals when it is obviously better.

Also, instead of trying to do a LIFO or FIFO or LIFO/FIFO policy for choosing
what to compute after a tactical goal is reached, we now have a policy that
allows continuing from any node (not just the last or first).  In order for
this to behave well for a variety of use cases, we needed to introduce the
concept of "epochs" and "eras", which define what can be computed later.

Performance-wise, this remains pretty good.  It is often a little slower, but
not significantly so.  We avoided pathological scaling behaviors.

* Oops.  Fix failing test.

* This tiny change makes `order` behave very similarly to the old `order`.

Similar to the old one, this occasionally has some quirks.
I'm hopeful that with a little more tinkering, we may find a "middle way" that
combines the best of the old way and the current way (on master and this branch).

* Yet another revision to `order`.

This applies a few of the lessons we have learned recently:
  1. there needs to be a way to change path along dependents if a better path is found.
  2. there needs to be a robust way to decide what to do next after a tactical goal
     (node with no dependents) is reached.
  3. when deciding what to do next, we should ideally take into account how close
     a node is to being completed (i.e., not just consider the final goal).

To do all these, we partition dependents of a node based on a score that indicates how
much future work is associated with the node.  We use this score to determine whether
tactical goals should change.  When all tactical goals are reached (so `inner_stack`
and `inner_stacks` are empty), we continue with the dependents of completed nodes that
had the lowest partition score.  We keep performance respectable by (hopefully)
partitioning nodes into many small groups (recall that it's much better to sort many
small groups than one large group).

* Fix docstring and doctest of `graph_metrics`

* Oups!  Fix off by one mistake.

* Use `defaultdict(list)` for readability (and maybe performance)

A couple benchmarks run faster for me, but most are about the same.
Still, I think using `defaultdict` is more readable here.  Thanks @TomAugspurger!

* Add test for `order` switching to a better dependent.

This increases coverage.

* Keep track of nodes that are in an `inner_stack`.

This fixes the performance regression `order.OrderLinearFull.time_order_linear_full`.
Most other benchmark change less than 3% (some faster, some slower),
but a couple artificial ones are 7%-10% slower.
This is unsurprising, since we do a little bit more work per edge,
which allows us to sometimes do a little less work per edge
(and, as the regressed benchmark shows, sometimes do a lot less work overall!).

No noticeable difference in behavior in graphs I've looked at.

* Add comments.

* Improve `order` by putting more dependents on inner_stacks when it's safe too.

Previously, `order` only added one dependent to `inner_stack` when `inner_stack` is empty.
Now, it adds all the other dependents with the minimum partition key to `inner_stacks`.
Prior to this commit, the new changes did a lot better with this example

https://github.com/dask/dask/issues/5859#issuecomment-608422198

but it didn't quite pass the new test.  In general, this means that non-optimal behavior
can happen when `inner_stacks` gets processed while there are better options available
in `next_nodes`.  Typically, this should only result in a modest delay.  We could do
better if we pushed dependents into, say, a sorted heapq, but this would kill performance.

* Conditionally add many dependents to inner_stacks.

Adding some dependents some of the time to inner_stacks can be beneficial,
but can be deterimental at other times.  Add a heuristic to try to decide.
Added tests that exemplify this.

* oops.  forgot to include the tests

* Don't condense `next_nodes` if it is large.  This isn't a performance bottleneck.

Co-authored-by: Tom Augspurger <tom.w.augspurger@gmail.com>