remove support for delayed bins as a scalar
breaks when `density` is True, since `diff` doesn't work on an Array with NaN chunks due to the slicing necessary. I opted to not support it at all, since arrays with NaN chunks are likely annoying to work with in general. but we could easily support it by just raising an error if `density and isinstance(bins, Array) and bins.ndim == 0`.