Histogram: support lazy values for range/bins (another way) (#6252)
* simpler? implementation of lazy histogram inputs

* remove support for delayed bins as a scalar

breaks when `density` is True, since `diff` doesn't work on an Array with NaN chunks due to the slicing necessary. I opted to not support it at all, since arrays with NaN chunks are likely annoying to work with in general. but we could easily support it by just raising an error if `density and isinstance(bins, Array) and bins.ndim == 0`.

* Revert "remove support for delayed bins as a scalar"

This reverts commit 58da8fdba4734e523928a02d972a89f75ba32c2e.

* support delayed scalar bins, except for `density`

* simplify nbins

* clearer error message re: scalar bins

* - restrict Dask types to Array and Delayed
- try to check scalariness of Delayed bins

* fixup! simplify nbins

* fixup! clearer error message re: scalar bins