Drop old frame splitting in NumPy serialization (#3868)
* Drop old frame splitting in NumPy serialization

This is leftover from when Blosc compression was still being handled
with the NumPy serialization code path itself. These days frame
splitting already happens as part of serialization more generally before
compressing data. Also this cutoff is not relevant for when splitting
occurs as that is determined by a limit in the size of buffers Blosc can
handle, which `frame_split_size` already knows. So go ahead and drop
this special casing in NumPy serialization and rely on these commonly
shared code paths that all serialization passes through.

* Drop `merge_frames` in NumPy deserialization

Since the frames are not being split in serialization and other common
code paths to all deserialization already make sure frames match their
expected length, there should be no need to merge frames in NumPy
deserialization. So drop this as well.

* Unpack `frames` in NumPy deserialization

To simplify the code a bit and assert the number of frames is exactly 1,
unpack `frames` into a single `frame` variable.