Drop oversubscription logic
Since `concurrent.futures` already has the logic necessary to avoid
oversubscription and our logic for it doesn't really apply, go ahead and
drop the oversubscription logic from the local scheduler and rely on
`concurrent.futures`.

Note it's still possible to configure pools with specific numbers of
workers. This doesn't affect that. Instead we just submit all tasks that
are ready until we are out of ready tasks.