Fix blockwise concatenate for array with some dimension=1.
This is an attempt at trying to understand and fix issue #6340,
this part of the codebase is relatively complex, so it is likely that
the fix is incorrect/incomplete.

Right now when we look at whether chunks should be concatenated. If so
we avoid generating the extra dummy [0] which are of help with
broadcasting as otherwise multiple copies of the arrays would be
concatenated.