Serialize `memoryview`s with `shape` and `format` (#3947)
* Standardize frame coercion to `memoryview`

Ensure that `frames` are coerced to `memoryview`s of `uint8` type
whether there are one or multiple `frames`.

* Track `memoryview`'s `format` and `shape`

When serializing a `memoryview`, track its format and shape. Make sure
to reinstantiate the `memoryview` on the other end with that format and
shape.

It's worth noting that `memoryview`s don't actually allow `strides` to
be passed as well. So we skip this. As a result if a F-order
`memoryview` is needed, one will need to perform a copy to fix this.

* Raise if a `memoryview` with `object`s is given

Trying to turn a `memoryview` of `object`s into `bytes` doesn't work
particularly well as it will contain only the memory pointers of the
Python objects without their actual data. So they won't be serialized
correctly when going over to the receiver. There isn't a good
alternative solution in this case. So just raise.

* Test serializing `memoryview` of `object`s