Synchronize on non-trivial CUDA frame transmission (#3949)
* Simplify `send_frames` comprehension

Use the same trick we use in `recv_frames` to determine which frames
will be needed for sending.

* Sync when transmitting non-trivial CUDA frames

Previously we would synchronize when any CUDA frames were part of
transmission regardless of whether they would be transmitted or not.
However we know that trivial (`0`-length) frames are not transmitted. So
this change checks whether there are non-trivial CUDA frames that will
be transmitted. Only if non-trivial CUDA frames are found, does
synchronization occur.