Add simple chunk type registry and defer as appropriate
Adds a basic list-based registry of valid chunk types for within
dask.array.Array, with the ability to add otherwise unknown types with
dask.array.register_chunk_type. When a type *not* in this list or
defined/handled by Dask is encountered in dunder operations or NumPy
functions/ufuncs, return NotImplemented.