Use `merge_frames` with host memory only (#3971)
* Use `merge_frames` with `"dask"` and `"pickle"`

Currently `merge_frames` is only really designed to work with memory on
host. So limit calling `merge_frames` to cases where host memory is used
(as opposed to cases with device memory).

* Check for CUDA frames before merging

As the `header` may not have the serialization mechanism used indicated,
simply check for any frames on device. If any such frames are found,
skip `merge_frames`.