Use `memoryview` in `unpack_frames` (#3980)
* Minor whitespace adjustment

* Coerce input to `memoryview` in `unpack_frames`

Selecting out each frame from the input causes a copy, which increases
memory usage and slows down `unpack_frames`. To fix this, coerce the
input to a `memoryview`. This way slices into the `memoryview` only take
a view onto the underlying data, which is quite fast and doesn't result
in additional memory usage.