Track mutable frames (#4004)
* Flag mutable frames as part of serialization

* Just use `bytes` in `merge_frames`

Keeps logic simple here in exchange for potentially copying later if
this is writeable.

* Only copy mutable frames

* Mark all CUDA frames as non-writeable

We don't need to bother making these frames writeable as they will be
transferred to GPU anyways where they will then be writeable. As a
result this saves us a copy that we don't otherwise need.

* Test NumPy array preserves writeability

* Fix-up `merge_frames` test with `writeable` header

* Relax the `memoryview` requirement of frames

* Drop test for frames being `memoryview`s

* Assert `writeable` and `lengths` have same # items

* Optionally determine `writeable`

Only try to figure out `writeable` if it is not otherwise specified.
This is important for CUDA objects for example, which are not
`memoryview` coercible.

* Rename `m` to `w` for clarity

* Fix Pandas test name

* Test writing to Pandas Series after serialization

* Force Pandas serialized frames to be readonly

Make sure all frames are `bytes` as they are read-only. This should test
whether the deserialization logic is smart enough to coerce them back
into something that is writeable.

* Test `merge_frames` with other `writeable` values

* Use read-only frames for serialized NumPy array

* Use `list` to compare `writeable`

* Drop unneeded CUDA array interface check

As we are only using `"dask"` or `"pickle"` serialization when spilling
here, we can be confident that there will not be any CUDA frames. So
drop this unneeded check.

* Use a `bytearray` to join writeable frames

* Handle singleton frame case as well

* Handle fast-path at the beginning

As we now guarantee frames are readonly or writeable as appropriate when
splitting/merging frames in all cases, there is no need to run this code
in all cases. We simply need it for the fast path. So add it there.

* Fix-up readonly case in fast path

Make sure we force readonly frames to be readonly.

* Mark CUDA frames as neither readonly no writeable

This special case of `None` will bypass both cases.

* Fix-up logic for copying singleton frame

* Test a few more `merge_frames` cases

* Drop unneeded CuPy customization

* Always use `join` path when copying is needed

* Go back to `.extend()`

This is very slightly faster than `.append()`

```python
In [1]: %timeit L = []
17.4 ns ± 0.0668 ns per loop (mean ± std. dev. of 7 runs, 100000000 loops each)

In [2]: L2 = [5]

In [3]: %timeit L = []; L.extend(L2)
73.7 ns ± 1.72 ns per loop (mean ± std. dev. of 7 runs, 10000000 loops each)

In [4]: %timeit L = []; L.append(L2[0])
84.5 ns ± 0.211 ns per loop (mean ± std. dev. of 7 runs, 10000000 loops each)
```

* Skip assignment and return `out`

* One more test with `bytearray`s

* Add function to check whether frame is writeable

Try to coerce the frame through `memoryview` and check the `readonly`
attribute. If the frame cannot be coerced through `memoryview`, use
`None` to indicate the frame is ambiguous, which means we will bypass
any special handling of the frame.

* Use `is_writeable` to detect writeable frames

* Explain return results from `is_writeable`