Serialize builtin array (#4013)
* Implement serialization for builtin `array`s

Leverage `array.frombytes(...)` for efficient copying of data into the
`array` using the Python Buffer Protocol. This requires that we make
things look `bytes`-like. In particular it requires the `itemsize == 1`,
which we facilitate by either casting to `uint8` (`"B"`) or copying to
`bytes`. Both will meet the requirement and be easily consumed.

* Test builtin `array` serialization